+++ Aplicación/Mod_Ingrediente/TiposIngredientes/Quesos.cs
@@ -8,7 +8,7 @@ namespace Aplicacion
 {
     public class Queso : PorPeso
     {
-        public Queso(int codigo, string nombre, int cantidadminima, decimal precioKilo, decimal cantidad)
+        public Queso(int codigo, string nombre, decimal cantidadminima, decimal precioKilo, decimal cantidad)
         {
             Codigo = codigo;
             Nombre = nombre;
